requirements.txt
aioredis==2.0.1
amqp==5.1.1
anyio==3.6.1
async-timeout==4.0.2
autoflake==1.4
bcrypt==3.2.2
billiard==3.6.4.0
black==22.6.0
bson==0.5.10
cachetools==5.2.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
contourpy==1.0.7
cryptography==37.0.4
cycler==0.11.0
Deprecated==1.2.13
dnspython==2.2.1
docker==5.0.3
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.78.0
flower==1.2.0
fonttools==4.38.0
google-auth==2.8.0
greenlet==1.1.2
h11==0.13.0
humanize==4.3.0
idna==3.3
importlib-resources==5.12.0
isort==5.10.1
jwt==1.3.1
kiwisolver==1.4.4
kombu==5.2.4
kubernetes==24.2.0
loguru==0.6.0
matplotlib==3.7.1
minio==7.1.11
mongoengine==0.24.1
mypy-extensions==0.4.3
networkx==2.8.6
numpy==1.24.1
oauthlib==3.2.0
packaging==21.3
pandas==1.5.3
passlib==1.7.4
pathspec==0.9.0
Pillow==9.4.0
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.30
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.9.1
pyflakes==2.4.0
pymongo==4.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tomli==2.0.1
toposort==1.7
tornado==6.2
typing-extensions==4.2.0
urllib3==1.26.9
uvicorn==0.18.1
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.3.3
websockets==10.3
wrapt==1.14.1
zipp==3.15.0
retrying==1.3.4
lakefs_client==0.102.2
